# nextjs-frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/repos/{owner}/{repo}/status": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Get Repository Status",
        "description": "Check the indexing status of a repository.",
        "operationId": "get_repository_status",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/repos/{owner}/{repo}/index": {
      "post": {
        "tags": [
          "repository"
        ],
        "summary": "Index Repository",
        "description": "Start the indexing process for a repository.\nCreates a Stripe checkout session for payment.",
        "operationId": "index_repository",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/repos/{owner}/{repo}/docs": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Get Repository Docs",
        "description": "Get documentation for a repository.",
        "operationId": "get_repository_docs",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/repos/{owner}/{repo}/info": {
      "get": {
        "tags": [
          "repository"
        ],
        "summary": "Get Github Repo Info",
        "description": "Get basic information about a GitHub repository.",
        "operationId": "get_github_repo_info",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/docs/{owner}/{repo}/file": {
      "get": {
        "tags": [
          "documentation"
        ],
        "summary": "Get File Docs",
        "description": "Get documentation for a specific file in a repository.",
        "operationId": "get_file_docs",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDescription"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutResponse": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          }
        },
        "type": "object",
        "required": [
          "checkout_url"
        ],
        "title": "CheckoutResponse"
      },
      "DocsResponse": {
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/RepositoryInfo"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileDescription"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "repository",
          "files"
        ],
        "title": "DocsResponse"
      },
...
      "FileDescription": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "path",
          "description",
          "type"
        ],
        "title": "FileDescription"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepoStatus": {
        "type": "string",
        "enum": [
          "not_indexed",
          "pending",
          "indexed",
          "failed"
        ],
        "title": "RepoStatus"
      },
      "RepositoryInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch"
          },
          "stars": {
            "type": "integer",
            "title": "Stars"
          },
          "forks": {
            "type": "integer",
            "title": "Forks"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "owner": {
            "$ref": "#/components/schemas/RepositoryOwner"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "full_name",
          "default_branch",
          "stars",
          "forks",
          "created_at",
          "updated_at",
          "size",
          "owner"
        ],
        "title": "RepositoryInfo"
      },
      "RepositoryOwner": {
        "properties": {
          "login": {
            "type": "string",
            "title": "Login"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "avatar_url": {
            "type": "string",
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "login",
          "id",
          "avatar_url"
        ],
        "title": "RepositoryOwner"
      },
      "RepositoryStatusResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RepoStatus"
          },
          "file_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Count"
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RepositoryStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}